Refactoring: move evaluation of binary operations into EvaluationUtil

+++ src/Parser/Parser.cs
@@ -517,7 +517,7 @@ public class Parser
 
             case TokenType.FloatLiteral:
                 tokens.Advance();
-                return new LiteralExpression(ValueType.Float, new Value(token.Value!.ToDecimal()));
+                return new LiteralExpression(ValueType.Float, new Value((double)token.Value!.ToDecimal()));
 
             case TokenType.Input:
                 tokens.Advance();
